requirements.txt
haversine==2.7.0
python==3.9.13
pandas==1.4.4
folium==0.14.0
streamlit==1.20.0
streamlit-folium==0.11.1
plotly==5.9.0
plotly-express==0.4.1
numpy==1.21.5
matplotlib==3.5.2
pillow==9.2.0
inflection == 0.5.1